+++ ds-core/src/doublelinkedlists.c
@@ -6,20 +6,11 @@
 
 int _dlist_node_initialize(dlist_node_t **node)
 {
-    // dlist_node_t *_node = NULL;
-    // if (!*node)
-    // {
-    //     puts("test");
-    //     _node = malloc(sizeof(dlist_node_t));
-    //     if (!_node) return -1;    
-    // }
-    // else
-    // {
-    //     puts("nodes maybe are allocated");
-    // }
-    
     dlist_node_t *_node = malloc(sizeof(dlist_node_t));
-    if (!_node) return -1;    
+    if (!_node)
+    {
+        return -1;
+    }
 
     _node->next = NULL;
     _node->prev = NULL;
@@ -46,9 +37,13 @@ dlist_node_t *dlist_append(dlist_node_t **head, dlist_node_t *node)
 {
     dlist_node_t *tail = dlist_get_tail(head);
     if (!tail)
+    {
         *head = node;
+    }
     else
+    {
         tail->next = node;
+    }
     node->prev = tail;
     node->next = NULL;
     return node;
@@ -62,10 +57,16 @@ size_t dlist_get_value(const dlist_item_t *item)
 int dlist_item_new(const size_t *value, dlist_item_t **item)
 {
     dlist_item_t *_item = malloc(sizeof(dlist_item_t));
-    if (!_item) return -1;
-    
+    if (!_item)
+    {
+        return -1;
+    }
+
     int res = _dlist_node_initialize( (dlist_node_t **)&_item->nodes);
-    if (res != 0) return -1;
+    if (res != 0)
+    {
+        return -1;
+    }
 
     _item->value = *value;
     *item = _item;
@@ -75,10 +76,9 @@ int dlist_item_new(const size_t *value, dlist_item_t **item)
 int dlist_item_clear(dlist_item_t *item)
 {
     item->value = NULL;
-    // free(item->value);
     item->nodes.prev = NULL;
     item->nodes.next = NULL;
-    // free((dlist_node_t *)&item->nodes);
+    item = NULL;
     free(item);
     return 0;
 }
@@ -88,18 +88,19 @@ dlist_item_t *dlist_search_by_value(dlist_item_t **head, const size_t *value, in
 {
     dlist_item_t *current_item = *head;
     dlist_item_t *item = NULL;
-    int pos = 0;
+    int pos = -1;
 
     while (current_item)
     {
         if (current_item->value == *value)
         {
             item = current_item;
-            *position = pos;
+            break;
         }
         current_item = (dlist_item_t *) current_item->nodes.next;
         ++pos;
     }
+    *position = pos;
     
     return item;
 }
@@ -126,7 +127,10 @@ dlist_item_t *dlist_item_remove_by_value(const size_t *value, dlist_item_t **hea
 
     int value_position = -1;
     result_item = dlist_search_by_value(head, value, &value_position);
-    if(!result_item) return NULL;
+    if(!result_item)
+    {
+        return NULL;
+    }
 
     result_item->nodes.prev->next = result_item->nodes.next;
     result_item->nodes.next->prev = result_item->nodes.prev;
@@ -138,7 +142,10 @@ dlist_item_t *dlist_item_remove_by_value(const size_t *value, dlist_item_t **hea
 
 int dlist_print(const dlist_item_t *head)
 {
-    if (!head) return -1;
+    if (!head)
+    {
+        return -1;
+    }
     
     dlist_item_t *current_node = head;
     int position = 0;
@@ -147,10 +154,50 @@ int dlist_print(const dlist_item_t *head)
     while (current_node)
     {
         // printf("%d. %zu\n", position, current_node->value);
-        printf("%d. %d\n", position, current_node->value);
+        printf("%d. %zu\n", position, current_node->value);
         current_node = (dlist_item_t *)current_node->nodes.next;
         ++position;
     }
     
+    return 0;
+}
+
+int dlist_insert_before(dlist_node_t *new_node, dlist_node_t *after_node)
+{
+    if (!new_node)
+    {
+        return -1;
+    }
+
+    if (!after_node)
+    {
+        return -1;
+    }
+
+    new_node->prev = after_node->prev;
+    new_node->next = after_node;
+    after_node->prev->next = new_node;
+    after_node->prev = new_node;
+
+    return 0;
+}
+
+int dlist_insert_after(dlist_node_t *new_node, dlist_node_t *previous_node)
+{
+    if (!new_node)
+    {
+        return -1;
+    }
+
+    if (!previous_node)
+    {
+        return -1;
+    }
+
+    new_node->prev = previous_node;
+    new_node->next = previous_node->next;
+    previous_node->next->prev = new_node;
+    previous_node->next = new_node;
+
     return 0;
 }